Working with confusion matrixs
Working with confusion matrixs and confused myself, am having my positive and negatives flipped. seems like my '0' is being classified as positive and my '1' being classified as negative.  Will look into it why this maybe, my thinking is it has to do with the data type for the predicted value y (target) maybe having it labeled as binary or classification would fix the issue.